+++ src/components/Plugins/LoadSave/PDBUrl.ts
@@ -4,8 +4,9 @@ import { IgnorePlugin } from "webpack";
 export class PDBUrl extends Parent.LoadSaveParent {
     private urlOrPDB;
     private environmentSelect: HTMLElement;
-    private shadowsObj;
-    private hardwareScalingObj;
+
+    public pluginTitle = "PDB";
+    public pluginSlug = "pdb-url";
 
     /**
      * Provides html describing the user interface (bootstrap4). Children
@@ -14,11 +15,11 @@ export class PDBUrl extends Parent.LoadSaveParent {
      */
     public userInterface(): string {
         return `
-            <h5>New Scene</h5>
+            <h5>Load PDB File</h5>
             <div class="row">
                 <div class="col-md-12">
                     <div class="form-group">
-                        <label for="exampleInputEmail1">
+                        <label for="pdbid-label">
                             Enter a PDB ID (e.g., <a class="link-sim" href="#" data-href="1XDN">1XDN</a>)
                             or PDB-/SDF-file URL (e.g., <a class="link-sim" href="#" data-href="https://files.rcsb.org/view/1XDN.pdb">https://files.rcsb.org/.../1XDN.pdb</a>,
                             <a class="link-sim" href="#" data-href="../nanokid.sdf" data-copy="nanokid.sdf">nanokid.sdf</a>).
@@ -50,45 +51,12 @@ export class PDBUrl extends Parent.LoadSaveParent {
                     </div>
                 </div>
             </div> -->
-            <hr class="mt-3 mb-4" style="border-top:0;" />
-            <h5>Graphics Settings</h5>
+            <!-- <hr class="mt-3 mb-4" style="border-top:0;" /> -->
             <!-- <div class="alert alert-warning enhance-resolution-container" style="display:none;">
                 Your browser does not support WebXR. ProteinVR will try to use the
                 older WebVR API instead.
-            </div> -->
-            <div class="row">
-                <div class="col-md-12">
-                    <div class="form-check">
-                        <label for="shadows">
-                            Molecular shadows can slow navigation and may crash some
-                            mobile browsers.
-                        </label>
-                        <div id="shadows">
-                            <input style="margin-left: 0;" type="checkbox" class="form-check-input" id="molecular-shadows">
-                            <label style="padding-left: 1.5rem;" class="form-check-label" for="molecular-shadows">Allow Molecules to Cast Subtle Shadows</label>
-                            <p style="height:5px;">&nbsp;</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row enhance-resolution-container" style="display:none;">
-                <div class="col-md-12">
-                    <div class="form-check">
-                        <label for="enhance-resolution">
-                            <!-- can cause problems on some headsets.--><!-- (e.g.,
-                            graphics in only one eye, left-eye field of view impinges
-                            on right, etc.).--> Try disabling enhanced VR resolution
-                            if ProteinVR sends incongruous images to each eye in VR mode.
-                            <!-- VR breaks on your device. -->
-                        </label>
-                        <div id="enhance-resolution">
-                            <input style="margin-left: 0;" type="checkbox" class="form-check-input" id="set-hardware-scaling-level">
-                            <label style="padding-left: 1.5rem;" class="form-check-label" checked for="set-hardware-scaling-level">Enhance Resolution in VR Mode (Recommended)</label>
-                            <p style="height:5px;">&nbsp;</p>
-                        </div>
-                    </div>
-                </div>
-            </div>`;
+            </div> -->` +
+            this._userInterfaceShadowsHardwareScaling();
     }
 
     /**
@@ -150,8 +118,6 @@ export class PDBUrl extends Parent.LoadSaveParent {
             this.environmentSelect.innerHTML = html;
 
             // Get jquery objects for the form elements.
-            this.shadowsObj = document.getElementById("molecular-shadows");
-            this.hardwareScalingObj = document.getElementById("set-hardware-scaling-level");
             var submitButton = document.getElementById("submit");
 
             // Set the values of those form elements if they are in local
@@ -159,33 +125,12 @@ export class PDBUrl extends Parent.LoadSaveParent {
             // if (localStorage.getItem("url") !== null) {
             //     urlOrPDB.val(localStorage.getItem("url"));
             // }
-            if (localStorage.getItem("shadows") !== null) {
-                // @ts-ignore
-                this.shadowsObj.checked = localStorage.getItem("shadows") === "true";
-            }
-
-            if (localStorage.getItem("hardwareScaling") !== null) {
-                // @ts-ignore
-                this.hardwareScalingObj.checked = localStorage.getItem("hardwareScaling") === "true";
-            } else {
-                // @ts-ignore
-                this.hardwareScalingObj.checked = true;  // default
-            }
-
-            if (window["webXRPolyfill"]["nativeWebXR"] !== true) {
-                var divs = document.getElementsByClassName("enhance-resolution-container");
-                for (var i = 0; i < divs.length; i++) {
-                    var div = divs[i];
-                    // @ts-ignore
-                    div.style.display = "inline-block";
-                }
-            }
 
             this.urlOrPDB.focus();
 
             // @ts-ignore
             submitButton.onclick = () => {
-                this.startLoadSave();
+                this._startLoadSave();
             };
 
             this.urlOrPDB.onkeypress = (e) => {
@@ -231,23 +176,15 @@ export class PDBUrl extends Parent.LoadSaveParent {
         var option = this.environmentSelect.options[this.environmentSelect.selectedIndex]; // .find("option:selected");
         var environ = option.value;
 
-        // @ts-ignore
-        var shadows = this.shadowsObj.checked;
-
-        // @ts-ignore
-        var hardwareScaling = this.hardwareScalingObj.checked;
-
         // Save them so they are the same when you reload. Decided
         // not to save url for security reasons (could be
         // proprietary).
         // localStorage.setItem('url', url);
         localStorage.setItem('environ', environ);
-        localStorage.setItem('shadows', shadows.toString());
-        localStorage.setItem('hardwareScaling', hardwareScaling.toString());
 
         // Construct the redirect url and redirect.
         let curUrl = window.location.protocol + "//" + window.location.host + window.location.pathname;
-        let newUrl = curUrl + "?s=" + url + "&e=" + environ + "&sh=" + shadows.toString() + "&hs=" + hardwareScaling.toString();
+        let newUrl = curUrl + "?s=" + url + "&e=" + environ + "&sh=" + this._shadowsObj.checked.toString() + "&hs=" + this._hardwareScalingObj.checked.toString();
         window.location.href = newUrl;
     }
 }