+++ CHANGES.md
@@ -1,6 +1,15 @@
 CHANGES
 =======
 
+WIP
+---
+
+* When reloading a saved ProteinVR URL, the molecular representations are now
+  loaded in the same, original order (in modern browsers).
+* Added proper citation (PLoS Comp Bio).
+* Improved error handling (e.g., warns when trying to load HTTP content over
+  HTTPS).
+
 1.0.4
 -----